+++ eventregistry/tests/TestQueryEventsComplex.py
@@ -1,6 +1,6 @@
 ﻿import unittest
 from eventregistry import *
-from DataValidator import DataValidator
+from .DataValidator import DataValidator
 
 class TestQueryEventsComplex(DataValidator):