+++ eventregistry/tests/TestQueryPaging.py
@@ -1,7 +1,7 @@
 from __future__ import print_function
 import unittest, math, random
 from eventregistry import *
-from DataValidator import DataValidator
+from .DataValidator import DataValidator
 from eventregistryadmin import EventRegistryAdmin