+++ eventregistry/tests/TestReturnInfo.py
@@ -3,7 +3,7 @@ test that setting return flags to false actually returns just the minimum set of
 """
 import unittest
 from eventregistry import *
-from DataValidator import DataValidator
+from .DataValidator import DataValidator
 
 class TestReturnInfo(DataValidator):
 
@@ -12,13 +12,13 @@ class TestReturnInfo(DataValidator):
             articleInfo=ArticleInfoFlags(bodyLen=0, basicInfo=False, title=False, body=False, url=False, eventUri=False, concepts=False, categories=False, links=False, videos=False, image=False, socialScore=False, sentiment=False, location=False, dates=False, extractedDates=False, duplicateList=False, originalArticle=False, storyUri=False),
             eventInfo=EventInfoFlags(title=False, summary=False, articleCounts=False, concepts=False, categories=False, location=False, date=False, commonDates=False, infoArticle=False, stories=False, socialScore=False, imageCount=0),
             sourceInfo = SourceInfoFlags(title = False, description = False,location = False,ranking = False,image = False,articleCount = False,socialMedia = False,sourceGroups = False),
-            conceptInfo=ConceptInfoFlags(type="concepts", lang="slv", label=False, synonyms=False, image=False, description=False, conceptClassMembership=False, conceptClassMembershipFull=False, trendingScore=False, trendingHistory=False, totalCount=False, trendingSource="news", maxConceptsPerType=20),
+            conceptInfo=ConceptInfoFlags(type="concepts", lang="slv", label=False, synonyms=False, image=False, description=False, conceptClassMembership=False, conceptClassMembershipFull=False, totalCount=False, trendingSource="news", maxConceptsPerType=20),
             locationInfo=LocationInfoFlags(label=False, wikiUri=False, geoNamesId=False, population=False, geoLocation=False, countryArea=False, countryDetails=False, countryContinent=False, placeFeatureCode=False, placeCountry=False),
             storyInfo = StoryInfoFlags(basicStats = False,location = False,date = False,title = False,summary = False,concepts = False,categories = False,medoidArticle = False,infoArticle = False,commonDates = False,socialScore = False,imageCount = 0),
             conceptClassInfo = ConceptClassInfoFlags(parentLabels = False))
 
     def ensureValidConcept(self, concept, testName):
-        for prop in ["label", "synonyms", "image", "trendingScore"]:
+        for prop in ["label", "synonyms", "image"]:
             self.assertFalse(prop in concept, "Property '%s' was not expected in concept for test %s" % (prop, testName))
         self.assertTrue(concept.get("type") in ["loc"], "Expected concept to be a location type, but got %s" % (concept.get("type")))
         if concept.get("location"):
@@ -38,7 +38,7 @@ class TestReturnInfo(DataValidator):
 
 
     def ensureValidCategory(self, category, testName):
-        for prop in ["parentUri", "trendingScore"]:
+        for prop in ["parentUri"]:
             self.assertFalse(prop in category, "Property '%s' was not expected in source for test %s" % (prop, testName))